requirements.txt
python-dotenv==1.0.0
flask==3.0.0
pymongo==4.6.0
redis==5.0.1